EleganceAI/requirements.txt
django==4.2.0
python-dotenv==1.0.1
genai==2.1.0
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.3.0
langchain-google-alloydb-pg==0.5.0
langchain-google-genai==1.0.10
langchain-huggingface==0.1.0
langchain-text-splitters==0.2.4
beautifulsoup4==4.12.3